Port 49c671e8, 11207ff1, 3abbe8b7, f49289a7
Fix build issues from the latest merge. These commits needed
porting:
49c671e8 odp: pktio: add pcapng capture capabilities
11207ff1 linux-gen: use inlined event type function
3abbe8b7 linux-gen: event: inline event_type function
f49289a7 linux-gen: buffer: remove buffer_inlines header file
