+++ platform/linux-dpdk/include/odp/api/plat/event_inlines.h
@@ -1,0 +1,1 @@
+../../../../../linux-generic/include/odp/api/plat/event_inlines.h
