+++ platform/linux-dpdk/include/odp_packet_internal.h
@@ -21,7 +21,6 @@ extern "C" {
 #include <odp_debug_internal.h>
 #include <odp/api/debug.h>
 #include <odp_buffer_internal.h>
-#include <odp_buffer_inlines.h>
 #include <odp_pool_internal.h>
 #include <odp/api/packet.h>
 #include <odp/api/plat/packet_inline_types.h>
