+++ platform/linux-dpdk/include/odp_packet_io_internal.h
@@ -126,6 +126,15 @@ struct pktio_entry {
 		odp_queue_t        queue;
 		odp_pktout_queue_t pktout;
 	} out_queue[PKTIO_MAX_QUEUES];
+
+	/* inotify instance for pcapng fifos */
+	struct {
+		enum {
+			PCAPNG_WR_STOP = 0,
+			PCAPNG_WR_PKT,
+		} state[PKTIO_MAX_QUEUES];
+		int fd[PKTIO_MAX_QUEUES];
+	} pcapng;
 };
 
 typedef union {
