+++ platform/linux-dpdk/m4/odp_pcapng.m4
@@ -1,0 +1,1 @@
+../../linux-generic/m4/odp_pcapng.m4
